The connection class make no any sense, It's a singleton and do not many thing in the context of application, because you can get once a instance and share across the routes using `use ($sharedConnection)` like syntax.
Create a config file returning a array with basic data to grant a connection with database.

On the `index.php` write the connection on the top of file and pass it to rest of routes. Also, all the exceptions can be get inside of `generator#clousure#connection` making logger a trivial task.

There are a lot of reasons to turn this closures on classes, but I think it can be as it is for now.